apps/backend/requirements-full.txt
# FastAPI and ASGI
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6

# Database
sqlalchemy[asyncio]==2.0.23
alembic==1.12.1
asyncpg==0.29.0
psycopg2-binary==2.9.9

# Vector Database
pgvector==0.2.4

# Redis and Job Queue
redis==5.0.1
arq==0.25.0

# Authentication and Security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6

# Rate Limiting
slowapi==0.1.9

# AI and ML
langchain>=0.0.350
langchain-openai>=0.0.2
langchain-anthropic>=0.0.1
openai>=1.6.1
anthropic>=0.8.0

# Git Operations
gitpython==3.1.40

# File Storage
boto3==1.34.0
minio==7.2.0

# Email
aiosmtplib==3.0.1
emails==0.6.0

# HTTP Client
httpx==0.25.2
aiohttp==3.9.1

# Configuration
pydantic==2.5.0
pydantic-settings==2.1.0

# Logging
loguru==0.7.2

# Utilities
tenacity==8.2.3
python-dotenv==1.0.0
pydantic==2.5.0

# Development and Testing
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-cov==4.1.0
httpx==0.25.2
black==23.11.0
ruff==0.1.6
mypy==1.7.1

# WebSocket
websockets==12.0

# Monitoring
sentry-sdk[fastapi]==1.38.0
